Refactor Dataframe Validation Logic
Refactors dataframe validation logic into 3 component files: column.py, dataframe.py, and variant.py.
This simplifies the validation structure and logically separates validation function based on the part of the df they operate on.
