+++ src/mavedb/lib/validation/constants/target.py
@@ -1,1 +1,2 @@
 valid_sequence_types = ["infer", "dna", "protein"]
+strict_valid_sequence_types = ["dna", "protein"]
